fixed protocol to receive RingSignPublicKeyType list not just byte array

+++ firmware/fsm.c
@@ -752,7 +752,7 @@ void fsm_msgRingSignMessage(RingSignMessage *msg) {
 	// this is for debugging
 
 	// print n
-	layoutNumber((uint32_t) msg->n, "n:");
+	layoutNumber(msg->n, "n:");
 	if (!protectButton(ButtonRequestType_ButtonRequest_PublicKey, true)) {
 		fsm_sendFailure(FailureType_Failure_ActionCancelled,
 				"Show public key cancelled");
